Add range to the page table walking code
Add a begin and end address to the page table code.  This will allow
individual addresses to be fetched, making it possible to direclty
fetch data from virtual addresses.

Also modify the MIPS code to pre-compute all the values it needs in
a separate function, making the processing more efficient for multiple
calls to the page walk code.