Fixup prstatus pr_pid if idle task is running.
Each CPU gets a prstatus note in the core.  However, the idle
thread always has pid 0, so if you have multiple CPUs running the
idle thread, you will get pid 0 multiple times.  This confuses gdb
and it looses all but one of the CPUs threads.  Instead, if the
idle thread is running, set the pid to -cpu to keep things sane.