Add handling for relocatable kernels
Kernels that are relocatable cannot be debugged as-is by gdb,
they require some changes.  This change adds an AT_ENTRY auxv
entry if an "entry" field is in the VMCOREINFO.  gdb can use
this and compare with the entry field in the vmlinux elf header
and relocate the core dump properly.

However, gdb will only do that with dynamic executables.  So
the program kdump-makedyn was added to convert an executable
from fixed address to dynamic type to fool gdb.

Kernel changes are needed for this, too, coming up later.

Signed-off-by: Corey Minyard <cminyard@mvista.com>