mips: Fix memory size calculations
The code was taking the task size and assuming the page table was
based off of that with the top removed.  However, that's apparently
not how it was done.  Instead, the entire page table's space was
used.  So, for instance, if the task size was 40 bits and the
page table span was 43 bits, it would use 43 bits of the VA for
addressing.

So calculate the page table span and use that for chopping off
the top bits and the bits that index into the page table.

Signed-off-by: Corey Minyard <cminyard@mvista.com>